Update EF models, update services and controllers

+++ BlogManagement/Controllers/AdminController.cs
@@ -19,31 +19,47 @@ namespace BlogManagement.Controllers
 
         [HttpGet]
         [Route("Blog")]
-        public async Task<IEnumerable<Blog>> GetBlogs() => await _blogService.GetAllBlogs();
-        
+        public async Task<ActionResult<IEnumerable<Blog>>> GetBlogs()
+        {
+            var result = await _blogService.GetAllBlogs();
+            return Ok(result);
+        }
+
         [HttpGet]
         [Route("Blog/{blogId}")]
-        public async Task<Blog> GetBlog(int blogId) => await _blogService.GetBlog(blogId);
+        public async Task<ActionResult<Blog>> GetBlog(int blogId)
+        {
+            var result = await _blogService.GetBlog(blogId);
+            if (result == null) return NotFound(blogId);
+            return Ok(result);
+        }
 
         [HttpPost]
         [Route("Blog")]
-        public async Task AddBlog([FromQuery] string blogName)
+        public async Task<ActionResult<Blog>> AddBlog([FromQuery] string blogName)
         {
-            await _blogService.AddBlog(blogName);
+            var result = await _blogService.AddBlog(blogName);
+            return Ok(result);
         }
         
         [HttpPut]
         [Route("Blog/{blogId}")]
-        public async Task UpdateBlog([FromRoute]int blogId, [FromQuery]string blogName)
+        public async Task<ActionResult<Blog>> UpdateBlog([FromRoute]int blogId, [FromQuery]string blogName)
         {
-            await _blogService.UpdateBlog(blogId, blogName);
+            var existingBlog = await _blogService.GetBlog(blogId);
+            if (existingBlog == null) return BadRequest($"Blog with id {blogId} is not exist");
+            var result = await _blogService.UpdateBlog(blogId, blogName);
+            return Ok(result);
         }
         
         [HttpDelete]
         [Route("Blog/{blogId}")]
-        public async Task RemoveBlog([FromRoute]int blogId)
+        public async Task<ActionResult> RemoveBlog([FromRoute]int blogId)
         {
+            var existingBlog = await _blogService.GetBlog(blogId);
+            if (existingBlog == null) return BadRequest($"Blog with id {blogId} is not exist");
             await _blogService.RemoveBlog(blogId);
+            return Ok();
         }
     }
 }
